# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WhyQuiet API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/batches": {
      "get": {
        "summary": "List Batches",
        "operationId": "list_batches_api_batches_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Batch"
                  },
                  "title": "Response List Batches Api Batches Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Batch",
        "operationId": "create_batch_api_batches_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Batch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/batches/{id}/approve": {
      "post": {
        "summary": "Approve Batch",
        "operationId": "approve_batch_api_batches__id__approve_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Batch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/batches/{id}/reject": {
      "post": {
        "summary": "Reject Batch",
        "operationId": "reject_batch_api_batches__id__reject_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Batch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/batches/{id}/export": {
      "get": {
        "summary": "Export Batch",
        "operationId": "export_batch_api_batches__id__export_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/profile": {
      "post": {
        "summary": "Profile",
        "operationId": "profile_api_profile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/triage": {
      "post": {
        "summary": "Triage",
        "operationId": "triage_api_triage_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/explain": {
      "post": {
        "summary": "Explain",
        "operationId": "explain_api_explain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/refuse": {
      "post": {
        "summary": "Refuse",
        "operationId": "refuse_api_refuse_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefuseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefuseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Batch": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "cause": {
            "$ref": "#/components/schemas/CauseFamily"
          },
          "remedy_code": {
            "type": "string",
            "title": "Remedy Code"
          },
          "unit_cost_bdt": {
            "type": "number",
            "title": "Unit Cost Bdt"
          },
          "wallet_count": {
            "type": "integer",
            "title": "Wallet Count"
          },
          "status": {
            "$ref": "#/components/schemas/BatchStatus"
          },
          "proposed_by": {
            "type": "string",
            "title": "Proposed By"
          },
          "decided_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decision_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Note"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "cause",
          "remedy_code",
          "unit_cost_bdt",
          "wallet_count",
          "status",
          "proposed_by",
          "created_at"
        ],
        "title": "Batch"
      },
      "BatchStatus": {
        "type": "string",
        "enum": [
          "proposed",
          "approved",
          "rejected"
        ],
        "title": "BatchStatus"
      },
      "CauseFamily": {
        "type": "string",
        "enum": [
          "job_exit",
          "migration",
          "solved_problem",
          "fee_shock",
          "supply_failure"
        ],
        "title": "CauseFamily"
      },
      "CreateBatchRequest": {
        "properties": {
          "cause": {
            "$ref": "#/components/schemas/CauseFamily"
          },
          "wallet_ids": {
            "items": {
              "type": "string",
              "pattern": "^W-[0-9A-Z]{6}$"
            },
            "type": "array",
            "maxItems": 1000,
            "minItems": 1,
            "title": "Wallet Ids"
          }
        },
        "type": "object",
        "required": [
          "cause",
          "wallet_ids"
        ],
        "title": "CreateBatchRequest"
      },
      "DecisionRequest": {
        "properties": {
          "note": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "note"
        ],
        "title": "DecisionRequest"
      },
      "ExplainRequest": {
        "properties": {
          "wallet_id": {
            "type": "string",
            "title": "Wallet Id"
          }
        },
        "type": "object",
        "required": [
          "wallet_id"
        ],
        "title": "ExplainRequest"
      },
      "ExportBatchResponse": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "cause": {
            "$ref": "#/components/schemas/CauseFamily"
          },
          "remedy_code": {
            "type": "string",
            "title": "Remedy Code"
          },
          "wallet_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Wallet Ids"
          },
          "cost_bdt": {
            "type": "number",
            "title": "Cost Bdt"
          },
          "approved_by": {
            "type": "string",
            "title": "Approved By"
          },
          "approved_at": {
            "type": "string",
            "title": "Approved At"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "cause",
          "remedy_code",
          "wallet_ids",
          "cost_bdt",
          "approved_by",
          "approved_at"
        ],
        "title": "ExportBatchResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "store": {
            "$ref": "#/components/schemas/StoreKind"
          },
          "stub": {
            "type": "boolean",
            "title": "Stub",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "status",
          "version",
          "store"
        ],
        "title": "HealthResponse"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "LoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "user_id",
          "role"
        ],
        "title": "LoginResponse"
      },
      "ProfileRequest": {
        "properties": {
          "wallet_id": {
            "type": "string",
            "title": "Wallet Id"
          }
        },
        "type": "object",
        "required": [
          "wallet_id"
        ],
        "title": "ProfileRequest"
      },
      "Reason": {
        "properties": {
          "feature": {
            "type": "string",
            "title": "Feature"
          },
          "contribution": {
            "type": "number",
            "title": "Contribution"
          },
          "direction": {
            "type": "string",
            "title": "Direction",
            "description": "'toward' or 'against' the attributed cause"
          }
        },
        "type": "object",
        "required": [
          "feature",
          "contribution",
          "direction"
        ],
        "title": "Reason"
      },
      "RefuseRequest": {
        "properties": {
          "wallet_id": {
            "type": "string",
            "title": "Wallet Id"
          }
        },
        "type": "object",
        "required": [
          "wallet_id"
        ],
        "title": "RefuseRequest"
      },
      "RefuseResponse": {
        "properties": {
          "verdict": {
            "$ref": "#/components/schemas/Verdict"
          },
          "cause": {
            "type": "null",
            "title": "Cause"
          },
          "posterior": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Posterior"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/Reason"
            },
            "type": "array",
            "title": "Reasons"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assumptions"
          },
          "generated_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated Text"
          },
          "refusal_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Refusal Reasons"
          },
          "stub": {
            "type": "boolean",
            "title": "Stub",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "verdict",
          "posterior",
          "reasons",
          "assumptions",
          "generated_text",
          "refusal_reasons"
        ],
        "title": "RefuseResponse"
      },
      "StoreKind": {
        "type": "string",
        "enum": [
          "local",
          "supabase"
        ],
        "title": "StoreKind"
      },
      "TriageRequest": {
        "properties": {
          "wallet_id": {
            "type": "string",
            "title": "Wallet Id"
          }
        },
        "type": "object",
        "required": [
          "wallet_id"
        ],
        "title": "TriageRequest"
      },
      "TriageResponse": {
        "properties": {
          "verdict": {
            "$ref": "#/components/schemas/Verdict"
          },
          "cause": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CauseFamily"
              },
              {
                "type": "null"
              }
            ]
          },
          "posterior": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Posterior"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/Reason"
            },
            "type": "array",
            "title": "Reasons"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assumptions"
          },
          "generated_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated Text"
          },
          "stub": {
            "type": "boolean",
            "title": "Stub",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "verdict",
          "cause",
          "posterior",
          "reasons",
          "assumptions",
          "generated_text"
        ],
        "title": "TriageResponse"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "analyst",
          "approver"
        ],
        "title": "UserRole"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Verdict": {
        "type": "string",
        "enum": [
          "attributed",
          "refused"
        ],
        "title": "Verdict"
      }
    }
  }
}